bgpv1: Split reconcilers into individual files
Splits recocnilers and their tests from a single source file to
a file per reconciler to avoid unnecessary code conflicts.

This change does not touch the logic of any reconciler, just
moves the code into newly created source files as it is.
